Create Error redirection page
Created a page that alerts user of an illegal action and prompts the user to go back to a different page to continue their experience. Example usage is to show an error when a user attempts to check out while not logged in, page informs user of that and gives a button for the user to go to the login page.